Moved cli to taskservice package (#40)
