+++ src/main/java/de/themorpheus/edu/taskservice/cli/commands/HelpCommand.java
@@ -1,9 +1,9 @@
-package de.themorpheus.edu.cli.commands;
+package de.themorpheus.edu.taskservice.cli.commands;
 
 import com.beust.jcommander.JCommander;
 import com.beust.jcommander.Parameters;
 
-import de.themorpheus.edu.cli.Options;
+import de.themorpheus.edu.taskservice.cli.Options;
 
 @Parameters(commandDescription = "Shows the help/usage page")
 public class HelpCommand {
